+++ src/validate/validate.schema.ts
@@ -611,6 +611,26 @@ export const profileStatusSchema: JSONSchema7 = {
   ...isNotEmpty('status'),
 };
 
+export const updateMessageSchema: JSONSchema7 = {
+  $id: v4(),
+  type: 'object',
+  properties: {
+    number: { type: 'string' },
+    text: { type: 'string' },
+    key: {
+      type: 'object',
+      properties: {
+        id: { type: 'string' },
+        remoteJid: { type: 'string' },
+        fromMe: { type: 'boolean', enum: [true, false] },
+      },
+      required: ['id', 'fromMe', 'remoteJid'],
+      ...isNotEmpty('id', 'remoteJid'),
+    },
+  },
+  ...isNotEmpty('number', 'text', 'key'),
+};
+
 export const profilePictureSchema: JSONSchema7 = {
   $id: v4(),
   type: 'object',
@@ -1127,7 +1147,18 @@ export const proxySchema: JSONSchema7 = {
   type: 'object',
   properties: {
     enabled: { type: 'boolean', enum: [true, false] },
-    proxy: { type: 'string' },
+    proxy: {
+      type: 'object',
+      properties: {
+        host: { type: 'string' },
+        port: { type: 'string' },
+        protocol: { type: 'string' },
+        username: { type: 'string' },
+        password: { type: 'string' },
+      },
+      required: ['host', 'port', 'protocol'],
+      ...isNotEmpty('host', 'port', 'protocol'),
+    },
   },
   required: ['enabled', 'proxy'],
   ...isNotEmpty('enabled', 'proxy'),